+++ R/LDpop.R
@@ -9,6 +9,9 @@
 #' @param r2d either "r2" for LD R2 or "d" for LD D', default = "r2"
 #' @param token LDlink provided user token, default = NULL, register for token at  \url{https://ldlink.nci.nih.gov/?tab=apiaccess}
 #' @param file Optional character string naming a path and file for saving results.  If file = FALSE, no file will be generated, default = FALSE.
+#' @param genome_build Choose between one of the three options...`grch37` for genome build GRCh37 (hg19),
+#' `grch38` for GRCh38 (hg38), or `grch38_high_coverage` for GRCh38 High Coverage (hg38) 1000 Genome Project
+#' data sets.  Default is GRCh37 (hg19).
 #'
 #' @return a data frame
 #' @importFrom httr GET content stop_for_status http_error
@@ -21,20 +24,28 @@
 #'                token = Sys.getenv("LDLINK_TOKEN"))
 #'              }
 #'
-LDpop <- function(var1, var2, pop = "CEU", r2d="r2", token=NULL, file = FALSE) {
+LDpop <- function(var1,
+                  var2,
+                  pop = "CEU",
+                  r2d="r2",
+                  token=NULL,
+                  file = FALSE,
+                  genome_build = "grch37") {
 
 LD_config <- list(ldpop_url="https://ldlink.nci.nih.gov/LDlinkRest/ldpop",
-                    avail_pop=c("YRI","LWK","GWD","MSL","ESN","ASW","ACB",
+                  avail_pop=c("YRI","LWK","GWD","MSL","ESN","ASW","ACB",
                                 "MXL","PUR","CLM","PEL","CHB","JPT","CHS",
                                 "CDX","KHV","CEU","TSI","FIN","GBR","IBS",
                                 "GIH","PJL","BEB","STU","ITU",
                                 "ALL", "AFR", "AMR", "EAS", "EUR", "SAS"),
-                    avail_ld=c("r2", "d"))
+                  avail_ld=c("r2", "d"),
+       avail_genome_build = c("grch37", "grch38", "grch38_high_coverage"))
 
 
   url <- LD_config[["ldpop_url"]]
   avail_pop <- LD_config[["avail_pop"]]
   avail_ld <- LD_config[["avail_ld"]]
+  avail_genome_build <- LD_config[["avail_genome_build"]]
 
 # ensure file option is a character string
 file <- as.character(file)
@@ -82,6 +93,15 @@ file <- as.character(file)
     stop("Enter valid access token. Please register using the LDlink API Access tab: https://ldlink.nci.nih.gov/?tab=apiaccess")
   }
 
+  # Ensure input for 'genome_build' is valid.
+  if(length(genome_build) > 1) {
+    stop("Invalid input.  Please choose only one available genome build.")
+  }
+
+  if(!(all(genome_build %in% avail_genome_build))) {
+    stop("Not an available genome build.")
+  }
+
 # Request body
 # snps_to_upload <- paste(unlist(snps), collapse = "%0A")
 pop_to_upload <- paste(unlist(pop), collapse = "%2B")
@@ -89,12 +109,13 @@ body <- list(paste("var1=", var1, sep=""),
              paste("var2=", var2, sep=""),
              paste("pop=", pop_to_upload, sep=""),
              paste("r2_d=", r2d, sep=""),
+             paste("genome_build=", genome_build, sep=""),
              paste("token=", token, sep=""))
 
 # URL query string
 url_str <- paste(url, "?", paste(unlist(body), collapse = "&"), sep="")
 
-# before 'POST command', check if LDlink server is up and accessible...
+# before 'GET' command', check if LDlink server is up and accessible...
 # if server is down pkg should fail gracefully with informative message (not error)
 if (httr::http_error(url)) { # if server is down use message (and not an error)
   message("The LDlink server is down or not accessible. Please try again later.")
@@ -123,13 +144,29 @@ names(data_out) <- gsub(x = names(data_out),
                       replacement = "_")
 
 # Check for error/warning in response data
-  if(grepl("error", data_out[nrow(data_out),1], ignore.case = TRUE)) {
-    stop(data_out[nrow(data_out),1])
-  }
+if(sum(grepl("error", data_out, ignore.case = TRUE), na.rm = TRUE)) {
+  # subset rows in data_out that contain text 'error'
+  error_msg <- subset(data_out, grepl("error", data_out[,1], ignore.case = TRUE))
 
-  if(grepl("warning", data_out[nrow(data_out),1], ignore.case = TRUE)) {
-    stop(data_out[nrow(data_out),1])
-  }
+  # delete any column names so that they don't go to output
+  names(error_msg) <- NULL
+
+  error_msg <- paste(error_msg, collapse = " ")
+
+  stop(error_msg)
+}
+
+if(sum(grepl("WARNING", data_out, ignore.case = TRUE), na.rm = TRUE)) {
+  # subset rows in data_out that contain text 'error'
+  warning_msg <- subset(data_out, grepl("WARNING", data_out[,1], ignore.case = TRUE))
+
+  # delete any column names so that they don't go to output
+  names(warning_msg) <- NULL
+
+  warning_msg <- paste(warning_msg, collapse = " ")
+
+  message(warning_msg)
+}
 
 # Evaluate 'file' option
   if (file == FALSE) {