+++ NEWS.md
@@ -1,3 +1,19 @@
+# LDlinkR 1.2.0
+* Issue #13 - Add new argument to `LDhap` (table_type) with four options for different types data output.
+* Issue #14 - Fix `LDhap` valid input check for number of input 'snps', line 94.
+* Add 'snps' to the WORDLIST
+* Update test-ldhap with additional tests
+* Update vignette with new `LDhap` usage examples with new argument.
+* Make LDhap able to write to file for both types of data returned by the new 'table_type' options.
+* Add feature 'genome_build' to support different genome builds, GRCh37(hg19), GRCh38(hg38), and GRCh38 High Coverage.
+* Improve/update error and warning message returned in the response data from the LDlink API server.
+* Fix CRAN note about 'LazyData' by removing line from Description file.
+* Update README file.
+* Add testthat tests for feature genome_build
+* Add three new Illumina chips to SNPchip
+* Add two new Affymetrix chips to SNPchip
+* Major update to vignette with new feature genome_build
+
 # LDlinkR 1.1.2
 * Submission date: 2021-02-19
 * Add link to LDlink web site documentation.